arduino python/requirements.txt
gTTS
SpeechRecognition
pyFirmata
python-engineio
pyttsx3
PyAudio